requirements.txt
pillow==8.0.0
